+++ examples/nanou_fireplace.rs
@@ -1,12 +1,13 @@
-use amber_light::ember::{self, Fireplace};
+
+use amber_light::fireplace::Fireplace;
 use canvas::digital_canvas::*;
 use nannou::prelude::*;
-use std::time;
+//use std::time;
 //use nannou_conrod::prelude::*;
 //use nannou_egui;
 
 struct Model {
-    main_window: WindowId,
+    _main_window: WindowId,
     my_canvas: DigitalCanvas,
     fireplace: Fireplace,
 }
@@ -18,7 +19,7 @@ fn main() {
         .run();
 }
 
-fn update(_app: &App, model: &mut Model, update: Update) {
+fn update(_app: &App, model: &mut Model, _update: Update) {
     //println!("{:?}", model.fireplace.settings.sigma);
     //model.fireplace.update_embers();
     //model.fireplace.find_heatmap();
@@ -64,7 +65,7 @@ fn model(app: &App) -> Model {
     //let egui = Egui::from_window(&window);
 
     let mut model = Model {
-        main_window: window,
+        _main_window: window,
         my_canvas: DigitalCanvas::new(),
         fireplace: Fireplace::new(),
     };
@@ -72,14 +73,14 @@ fn model(app: &App) -> Model {
     return model;
 }
 
-fn raw_window_event(_app: &App, model: &mut Model, event: &nannou::winit::event::WindowEvent) {
+fn raw_window_event(_app: &App, _model: &mut Model, _event: &nannou::winit::event::WindowEvent) {
     // Let egui handle things like keyboard and mouse input.
     //model.egui.handle_raw_event(event);
 }
 
-fn raw_ui_event(_app: &App, _model: &mut Model, _event: &nannou_conrod::RawWindowEvent) {}
+fn _raw_ui_event(_app: &App, _model: &mut Model, _event: &nannou_conrod::RawWindowEvent) {}
 
-fn key_pressed(app: &App, model: &mut Model, key: Key) {
+fn key_pressed(_app: &App, model: &mut Model, key: Key) {
     match key {
         Key::E => {
             model.fireplace.settings.ember_settings.sigma =
@@ -96,7 +97,7 @@ fn key_pressed(app: &App, model: &mut Model, key: Key) {
     }
 }
 
-fn ui_view(app: &App, model: &Model, frame: Frame) {}
+fn _ui_view(_app: &App, _model: &Model, _frame: Frame) {}
 
 fn view(app: &App, model: &Model, frame: Frame) {
     // Begin drawing
